+++ examples/authentication/src/main/java/cloud/stackit/sdk/authentication/examples/AuthenticationExample.java
@@ -1,6 +1,7 @@
 package cloud.stackit.sdk.authentication.examples;
 
 import cloud.stackit.sdk.core.config.CoreConfiguration;
+import cloud.stackit.sdk.core.exception.ApiException;
 import cloud.stackit.sdk.resourcemanager.api.ResourceManagerApi;
 import cloud.stackit.sdk.resourcemanager.model.ListOrganizationsResponse;
 import java.io.File;
@@ -8,15 +9,21 @@ import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.util.Scanner;
 
-class AuthenticationExample {
-	public static void main(String[] args) throws IOException {
-		///////////////////////////////////////////////////////
-		// Option 1: setting the paths to service account key (and private key) as configuration
-		///////////////////////////////////////////////////////
-		final String SERVICE_ACCOUNT_KEY_PATH = "/path/to/sa_key.json";
-		final String PRIVATE_KEY_PATH = "/path/to/private_key.pem";
-		final String SERVICE_ACCOUNT_MAIL = "name-1234@sa.stackit.cloud";
+final class AuthenticationExample {
 
+	private static final String SERVICE_ACCOUNT_KEY_PATH = "/path/to/sa_key.json";
+	private static final String PRIVATE_KEY_PATH = "/path/to/private_key.pem";
+	private static final String SERVICE_ACCOUNT_MAIL = "name-1234@sa.stackit.cloud";
+
+	private AuthenticationExample() {}
+
+	@SuppressWarnings({
+		"PMD.CyclomaticComplexity",
+		"PMD.SystemPrintln",
+		"PMD.AvoidThrowingRawExceptionTypes"
+	})
+	public static void main(String[] args) throws IOException {
+		/* OPTION 1: setting the paths to service account key (and private key) as configuration */
 		try {
 			ResourceManagerApi api =
 					new ResourceManagerApi(
@@ -29,17 +36,19 @@ class AuthenticationExample {
 			ListOrganizationsResponse response =
 					api.listOrganizations(null, SERVICE_ACCOUNT_MAIL, null, null, null);
 
-			System.out.println(response);
-		} catch (Exception e) {
+			System.out.println(response.toString());
+		} catch (ApiException | IOException e) {
 			throw new RuntimeException(e);
 		}
 
-		///////////////////////////////////////////////////////
-		// Option 2: setting the service account key (and private key) as configuration
-		///////////////////////////////////////////////////////
+		/*
+		 * OPTION 2: setting the service account key (and private key) as configuration
+		 *
+		 * */
 
-		// read key content from a file, in production you can also read it e.g. from STACKIT
-		// secrets manager, so it's only kept in-memory
+		/* read key content from a file, in production you can also read it
+		 * e.g. from STACKIT secrets manager, so it's only kept in-memory
+		 * */
 		String serviceAccountKeyPath = // replace it with the path to your service account key
 				"examples/authentication/src/main/java/cloud/stackit/sdk/authentication/examples/dummy_credentials/dummy-service-account-key.json";
 		File serviceAccountKeyFile = new File(serviceAccountKeyPath);
@@ -49,7 +58,8 @@ class AuthenticationExample {
 				serviceAccountKeyContent.append(myReader.nextLine());
 			}
 		} catch (FileNotFoundException e) {
-			throw new RuntimeException(e);
+			System.err.println("File not found: " + serviceAccountKeyPath);
+			return;
 		}
 
 		String privateKeyPath = // replace it with the path to your private key
@@ -61,7 +71,8 @@ class AuthenticationExample {
 				privateKeyContent.append(myReader.nextLine());
 			}
 		} catch (FileNotFoundException e) {
-			throw new RuntimeException(e);
+			System.err.println("File not found: " + privateKeyPath);
+			return;
 		}
 
 		String serviceAccountKey = serviceAccountKeyContent.toString();
@@ -79,35 +90,36 @@ class AuthenticationExample {
 			ListOrganizationsResponse response =
 					api.listOrganizations(null, SERVICE_ACCOUNT_MAIL, null, null, null);
 
-			System.out.println(response);
-		} catch (Exception e) {
+			System.out.println(response.toString());
+		} catch (ApiException | IOException e) {
 			throw new RuntimeException(e);
 		}
 
-		///////////////////////////////////////////////////////
-		// Option 3: setting the service account key (and private key) as environment variable
-		///////////////////////////////////////////////////////
-		// Set the service account key via environment variable:
-		// - STACKIT_SERVICE_ACCOUNT_KEY_PATH=/path/to/sa_key.json
-		// - STACKIT_SERVICE_ACCOUNT_KEY="<content of service account key>"
-		//
-		// If the private key is not included in the service account key, set also:
-		// - STACKIT_PRIVATE_KEY_PATH=/path/to/private_key.pem
-		// - STACKIT_PRIVATE_KEY="<content of private key>"
-		//
-		// If no environment variable is set, fallback to credentials file in
-		// "$HOME/.stackit/credentials.json".
-		// Can be overridden with the environment variable `STACKIT_CREDENTIALS_PATH`
-		// The credentials file must be a json:
-		//   {
-		//     "STACKIT_SERVICE_ACCOUNT_KEY_PATH": "path/to/sa_key.json",
-		//     "STACKIT_PRIVATE_KEY_PATH": "(OPTIONAL) when the private key isn't included in the
-		// Service Account key",
-		//     // Alternative:
-		//     "STACKIT_SERVICE_ACCOUNT_KEY": "<content of private key>",
-		//     "STACKIT_PRIVATE_KEY": "(OPTIONAL) when the private key isn't included in the Service
-		// Account key",
-		//   }
+		/*
+		 * OPTION 3: setting the service account key (and private key) as environment variable
+		 *
+		 * Set the service account key via environment variable:
+		 * - STACKIT_SERVICE_ACCOUNT_KEY_PATH=/path/to/sa_key.json
+		 * - STACKIT_SERVICE_ACCOUNT_KEY="<content of service account key>"
+		 *
+		 * If the private key is not included in the service account key, set also:
+		 * - STACKIT_PRIVATE_KEY_PATH=/path/to/private_key.pem
+		 * - STACKIT_PRIVATE_KEY="<content of private key>"
+		 *
+		 * If no environment variable is set, fallback to credentials file in
+		 * "$HOME/.stackit/credentials.json".
+		 * Can be overridden with the environment variable `STACKIT_CREDENTIALS_PATH`
+		 * The credentials file must be a json:
+		 *   {
+		 *     "STACKIT_SERVICE_ACCOUNT_KEY_PATH": "path/to/sa_key.json",
+		 *     "STACKIT_PRIVATE_KEY_PATH": "(OPTIONAL) when the private key isn't included in the
+		 * Service Account key",
+		 *     // Alternative:
+		 *     "STACKIT_SERVICE_ACCOUNT_KEY": "<content of private key>",
+		 *     "STACKIT_PRIVATE_KEY": "(OPTIONAL) when the private key isn't included in the Service
+		 * Account key",
+		 *   }
+		 * */
 		try {
 			ResourceManagerApi api = new ResourceManagerApi();
 
@@ -115,8 +127,8 @@ class AuthenticationExample {
 			ListOrganizationsResponse response =
 					api.listOrganizations(null, SERVICE_ACCOUNT_MAIL, null, null, null);
 
-			System.out.println(response);
-		} catch (Exception e) {
+			System.out.println(response.toString());
+		} catch (ApiException | IOException e) {
 			throw new RuntimeException(e);
 		}
 	}